OSPRH-17209: Change the default boot label.
A conflict can occur where storage area networks are utilized for
instance storage, where multiple instances might exist on the
underlying backend storage, which may conflict with the local disks.

For RHOSO, we want the the intended and expected disk to boot, and
*not* any other potential disk.

A such, we've changed the default label moving forward.

Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>